server/requirements.txt
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
django-filter==21.1
psycopg2==2.9.3
djoser==2.1.0
